refactor(core): split diagram extension into focused modules
Extract the two pure, DOM-free concerns from the 760-line diagram
extension into co-located modules:

- diagram-render.ts: the mermaid / graphviz lazy loaders, SVG rendering,
  initialization, theme resolution, and diagram-type helpers.
- diagram-markdown.ts: the fence-language mapping and the markdown-it
  fence renderer override.

diagram.ts keeps the node spec, options, and factory, and re-exports the
VizelDiagramType and GraphvizEngine types so the public surface is
unchanged.
